refactor: LoadingSpinner 컴포넌트 통합 및 size/text prop 지원 추가
